Fixed conflict and UI

+++ admin/category/category.php
@@ -32,13 +32,13 @@ require '../../connect.php';
                         <a class="nav-link" href="../index.php">Trang ngoài</a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link active fw-bold" href="category.php">Thể loại</a>
+                        <a class="nav-link active fw-bold" href="../category/category.php">Thể loại</a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link" href="author.php">Tác giả</a>
+                        <a class="nav-link" href="../author/author.php">Tác giả</a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link" href="article.php">Bài viết</a>
+                        <a class="nav-link" href="../article/article.php">Bài viết</a>
                     </li>
                 </ul>
                 </div>
@@ -62,7 +62,7 @@ require '../../connect.php';
                     </thead>
                     <tbody>
                         <?php 
-                            $query ="SELECT ma_tloai,ten_tloai FROM theloai";
+                            $query ="SELECT ma_tloai,ten_tloai FROM theloai ORDER BY ma_tloai DESC";
                             $statement = $conn->prepare($query);
                             $statement->execute();
                             
@@ -77,7 +77,7 @@ require '../../connect.php';
                                                 <a href="edit_category.php?id=<?= $row->ma_tloai; ?>"><i class="fa-solid fa-pen-to-square"></i></a>
                                             </td>
                                             <td>
-                                                <a href="delete_category.php?id=<?= $row->ma_tloai; ?>"><i class="fa-solid fa-trash"></i></a>
+                                                <a onclick="return confirm('Bạn có chắc chắn muốn xoá?')" href="delete_category.php?id=<?= $row->ma_tloai; ?>"><i class="fa-solid fa-trash"></i></a>
                                             </td>
                                         </tr>
                                     <?php
